move back button to a component
